moved shared exomol files in API, fix tests
