+++ radis/tools/line_survey.py
@@ -13,15 +13,15 @@ from warnings import warn
 
 import numpy as np
 
-from radis.db.classes import get_molecule, get_molecule_identifier
-from radis.io.cdsd import columns_4000 as cdsd4000columns
-from radis.io.cdsd import columns_cdsdhitemp as cdsdhitempcolumns
-from radis.io.hitran import (  # HITRAN_CLASS2,; HITRAN_CLASS3,; HITRAN_CLASS6,; HITRAN_CLASS7,; HITRAN_CLASS8,; HITRAN_CLASS9,; HITRAN_CLASS10,
+from radis.api.cdsdapi import columns_4000 as cdsd4000columns
+from radis.api.cdsdapi import columns_cdsdhitemp as cdsdhitempcolumns
+from radis.api.hitranapi import (  # HITRAN_CLASS2,; HITRAN_CLASS3,; HITRAN_CLASS6,; HITRAN_CLASS7,; HITRAN_CLASS8,; HITRAN_CLASS9,; HITRAN_CLASS10,
     HITRAN_CLASS1,
     HITRAN_CLASS4,
     HITRAN_CLASS5,
 )
-from radis.io.hitran import columns_2004 as hitrancolumns
+from radis.api.hitranapi import columns_2004 as hitrancolumns
+from radis.db.classes import get_molecule, get_molecule_identifier
 from radis.misc.basics import is_float
 from radis.misc.printer import superscript_map
 from radis.misc.utils import NotInstalled
